change BIGINT to INT

+++ example/all_test.go
@@ -14,7 +14,7 @@ import (
 func TestCreate(t *testing.T) {
 	t.Parallel()
 	assert.Equal(t,
-		`CREATE TABLE 'all' ( 'int' BIGINT PRIMARY KEY, 'string' VARCHAR(100) NOT NULL, 'bool' BOOLEAN, 'time' TIMESTAMP, 'select' BIGINT )`,
+		`CREATE TABLE 'all' ( 'int' INT PRIMARY KEY, 'string' VARCHAR(100) NOT NULL, 'bool' BOOLEAN, 'time' TIMESTAMP, 'select' INT )`,
 		new(aorm.ORM).Create().String(),
 	)
 }
